Fixes the interpretation of what a sentence is.
The definition of a sentence for usage by Emacs' editing commands was
more like a human definition of a paragraph. This makes it work better,
though there's still some edge cases at the ends of paragraphs.

Could fix those edge cases with some regex. Or putting two spaces after
each sentence, but that's... not a good reason to switch.